pyproject.toml
[project]
name = "audocare-ai-agent"
version = "0.1.0"
description = "AI voice agent for Audocare"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "livekit-agents[openai,elevenlabs,silero,turn-detector]~=1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "openai>=1.79.0",
    "python-dotenv>=1.1.0",
]
